+++ ubo_app/side_effects.py
@@ -167,6 +167,7 @@ def setup_side_effects() -> Subscriptions:
             {
                 'id': service.id,
                 'is_enabled': service.is_enabled,
+                'log_level': service.log_level,
                 'should_auto_restart': service.should_auto_restart,
             }
             for service in state.settings.services.values()